feat: use isBridge method (wip)
