+++ springwolf-core/src/main/java/io/github/stavshamir/springwolf/asyncapi/scanners/channels/operationdata/annotation/AsyncListenerAnnotationScanner.java
@@ -66,9 +66,9 @@ public class AsyncListenerAnnotationScanner extends AbstractOperationDataScanner
         log.debug("Scanning class \"{}\" for @\"{}\" annotated methods", type.getName(), annotationClass.getName());
 
         return Arrays.stream(type.getDeclaredMethods())
+                .filter(method -> !method.isBridge())
                 .filter(method -> method.isAnnotationPresent(annotationClass)
-                        || method.isAnnotationPresent(annotationClassRepeatable))
-                .filter(method -> !AsyncAnnotationScannerUtil.isMethodInherited(method, type));
+                        || method.isAnnotationPresent(annotationClassRepeatable));
     }
 
     private Stream<OperationData> toOperationData(Method method) {
