+++ springwolf-core/src/test/java/io/github/stavshamir/springwolf/asyncapi/scanners/channels/operationdata/annotation/AsyncPublisherAnnotationScannerIntegrationTest.java
@@ -17,6 +17,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
@@ -205,10 +207,11 @@ class AsyncPublisherAnnotationScannerIntegrationTest {
         assertThat(actualChannels).containsExactly(Map.entry("test-channel", expectedChannel));
     }
 
-    @Test
-    void scan_componentHasOnlyDeclaredMethods() {
+    @ParameterizedTest
+    @ValueSource(classes = {ClassImplementingInterface.class, ClassImplementingInterfaceWithAnnotation.class})
+    void scan_componentHasOnlyDeclaredMethods(Class<?> clazz) {
         // Given a class with a method, which is declared in a generic interface
-        setClassToScan(ClassImplementingInterface.class);
+        setClassToScan(clazz);
 
         // When scan is called
         Map<String, ChannelItem> actualChannels = channelScanner.scan();
@@ -276,7 +279,7 @@ class AsyncPublisherAnnotationScannerIntegrationTest {
         private void methodWithMultipleAnnotation(SimpleFoo payload) {}
     }
 
-    private static class ClassImplementingInterface implements AsyncAnnotationScannerUtilTest.ClassInterface<String> {
+    private static class ClassImplementingInterface implements ClassInterface<String> {
 
         @AsyncPublisher(
                 operation =
@@ -285,6 +288,25 @@ class AsyncPublisherAnnotationScannerIntegrationTest {
                                 description = "test channel operation description"))
         @Override
         public void methodFromInterface(String payload) {}
+    }
+
+    interface ClassInterface<T> {
+        void methodFromInterface(T payload);
+    }
+
+    private static class ClassImplementingInterfaceWithAnnotation implements ClassInterfaceWithAnnotation<String> {
+
+        @Override
+        public void methodFromInterface(String payload) {}
+    }
+
+    interface ClassInterfaceWithAnnotation<T> {
+        @AsyncPublisher(
+                operation =
+                        @AsyncOperation(
+                                channelName = "test-channel",
+                                description = "test channel operation description"))
+        void methodFromInterface(T payload);
     }
 
     @Data
